Update styles.css
Key Changes:
Background Image for "Listen to Reaktor4":

background-image: url('Reaktor4 The Cabin.jpeg');: This uses the image "Reaktor4 The Cabin.jpeg" as the background.
background-repeat: repeat;: The image will be tiled across the section to create a collage effect.
background-size: 300px;: This controls the size of the image in the collage. You can adjust the 300px value to make the tiles larger or smaller depending on your preferences.